fix: satisfy translation static analysis
